// src/modules/core/clusters/md-user/pages/user/types/ManageUserPage.types.ts
type Profile = {
  id_number: string
  full_name: string
}

type Werk = {
  name: string
}

type Sto = {
  name: string
}

type User = {
  id: string
  email: string | null
  is_verified: boolean
  status: string
  profile: Profile | null
  werk: Werk | null
  sto: Sto | null
}

type GetTableUsersMeta = {
  filter_count: number
}

export type GetTableUsersResponse = {
  data: User[]
  meta: GetTableUsersMeta
}